+++ systemimage/candidates.py
@@ -17,7 +17,7 @@
 
 __all__ = [
     'get_candidates',
-    'get_downloads',
+    'iter_path',
     ]
 
 
@@ -119,21 +119,20 @@ def get_candidates(index, build):
     return paths
 
 
-def get_downloads(winner):
-    """Return the list of files to download given the upgrade winner.
+def iter_path(winner):
+    """Iterate over all the file records for a given upgrade path.
 
     Image traversal will stop after the first `bootme` flag is seen, so the
     list of files to download may not include all the files in the upgrade
     candidate.
 
     :param winner: The list of images for the winning candidate.
-    :return: The list of file records describing the files to download.  Each
-        file record is a dictionary as described in the index.json file and
-        contains such information as the path to the signature file, the
-        file's checksum and size, and its order.
+    :return: A sequence of 2-tuples, where the first item is a "image
+        number", i.e. which image in the path of winning images this file
+        record belongs to, and the second item is the file record.
     """
-    for image in winner:
+    for n, image in enumerate(winner):
         for filerec in image.files:
-            yield filerec
+            yield (n, filerec)
         if getattr(image, 'bootme', False):
             break