+++ systemimage/tests/test_winner.py
@@ -24,7 +24,7 @@ import os
 import unittest
 
 from contextlib import ExitStack
-from systemimage.candidates import get_candidates, get_downloads
+from systemimage.candidates import get_candidates
 from systemimage.config import config
 from systemimage.download import get_files
 from systemimage.gpg import SignatureError